requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
Pillow>=10.0.0
python-dateutil>=2.8.2
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
djangorestframework-simplejwt